Submission Logic (#3)
* 1) More or less done with dataloader
2) updated models to be used with the new config dict
3) tested stuff in playgroud notebook
4) Added a couple of utility functions in utils including a augmentation
   visualization function
5) modified train to work with dataloader and seems to be working. Needs
   more polishing but more or less done.

* updated requirements

* added more comments, push batches to specific device before feeding it to the model

* Add Makefile

This file contains all the commands for making submissions.

* Add benchmark_src

This is example code for how to make submission. You can make this submission with the "make pack-benchmark" command.

* Add submission main.py

This is the file we will use for submissions. The predict function is where our model will work. A submission can be made with the "make pack-submission" command.

* Restructured project

* fixed .gitignore

Co-authored-by: Gautham Venkataraman <gauthsvenkat99@gmail.com>